+++ tasks/src/main/java/edu/pjwstk/tasks/application/deletetasknotification/DeleteTaskNotificationUseCaseImpl.java
@@ -1,7 +1,10 @@
 package edu.pjwstk.tasks.application.deletetasknotification;
 
+import edu.pjwstk.common.authApi.AuthApi;
+import edu.pjwstk.common.authApi.dto.CurrentUserDto;
 import edu.pjwstk.tasks.entity.TaskNotification;
 import edu.pjwstk.tasks.exception.TaskNotFoundException;
+import edu.pjwstk.tasks.exception.UnauthorizedTaskAccessException;
 import edu.pjwstk.tasks.repository.TaskNotificationRepository;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,19 +15,28 @@ import java.util.UUID;
 public class DeleteTaskNotificationUseCaseImpl implements DeleteTaskNotificationUseCase {
 
     private final TaskNotificationRepository taskNotificationRepository;
+    private final AuthApi currentUserProvider;
 
-    public DeleteTaskNotificationUseCaseImpl(TaskNotificationRepository taskNotificationRepository) {
+    public DeleteTaskNotificationUseCaseImpl(TaskNotificationRepository taskNotificationRepository, AuthApi currentUserProvider) {
         this.taskNotificationRepository = taskNotificationRepository;
+        this.currentUserProvider = currentUserProvider;
     }
 
     @Override
     @Transactional
     public void execute(UUID taskId, Integer taskNotificationId) {
+
+
         TaskNotification taskNotification = taskNotificationRepository
                 .findByIdAndTaskId(taskId, taskNotificationId)
                 .orElseThrow(() -> new TaskNotFoundException(
                         "Task with id " + taskId + " not found!"
                 ));
+
+        CurrentUserDto currentUserDto = currentUserProvider.getCurrentUser().orElseThrow();
+        if (currentUserDto.userId() != taskNotification.getTask().getUserId()) {
+            throw new UnauthorizedTaskAccessException("User is not authorized to delete notification for another user!");
+        }
 
         taskNotificationRepository.deleteByIdAndTaskId(taskId, taskNotificationId);
     }
